requirements.txt
pytest=7.2.1
requests=2.28.2
jsonschema=4.17.3
pyhamcrest=2.0.4